feat: Add zod validation middleware
Copied from https://dev.to/osalumense/validating-request-data-in-expressjs-using-zod-a-comprehensive-guide-3a0j

This is awesome because it means we can get useful validation error messages for every request.

There exists a package (https://github.com/Aquila169/zod-express-middleware) but it looks dead.
This implementation doesn't coerce the req.body passed to the route to the right type,
but it's not too big of a lift to parse the schema again in the route.
...
for now.