feat: Decouple user type from Kinde
I'm on the plane, and I don't have a good enough connection to auth with
Kinde. That's annoying, so let's do what we were supposed to do before
and make our own user type.

Also make a backdoor dev user that we can access with an env flag.
Probably should remove this for security when this goes to prod.