feat: Mock grade
Laying some pipe for eventual grading.

Next step is to implement it in scenarioEngine to actually do something.

Cursor brings up that having grade as a separate JSON object on the
ScenarioSession table might make future migrations easier when we want
to make Grade a table of its own so that its indexable for querying and
analytics. But for now, this is just fine.